# web/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "The Buzz Crew API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/leads": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "List Leads",
        "operationId": "listLeads",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "leads"
        ],
        "summary": "Submit a project inquiry (public)",
        "operationId": "createLead",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadCreated"
                }
              }
            }
          },
          "429": {
            "description": "Too many submissions from this visitor"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/{lead_id}": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "Get Lead",
        "operationId": "getLead",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "leads"
        ],
        "summary": "Update Status",
        "operationId": "updateLeadStatus",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lead Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "leads"
        ],
        "summary": "Delete Lead",
        "operationId": "deleteLead",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/{lead_id}/notes": {
      "post": {
        "tags": [
          "leads"
        ],
        "summary": "Add Note",
        "operationId": "addLeadNote",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lead Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewNote"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadNoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/newsletter": {
      "post": {
        "tags": [
          "newsletter"
        ],
        "summary": "Subscribe",
        "operationId": "subscribe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/subscribers/{subscriber_id}": {
      "delete": {
        "tags": [
          "newsletter"
        ],
        "summary": "Delete Subscriber",
        "operationId": "deleteSubscriber",
        "parameters": [
          {
            "name": "subscriber_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subscriber Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BudgetRange": {
        "type": "string",
        "enum": [
          "under_50k",
          "50k_150k",
          "150k_plus",
          "not_sure"
        ],
        "title": "BudgetRange"
      },
...
      "Country": {
        "type": "string",
        "enum": [
          "pakistan",
          "uae",
          "uk",
          "other"
        ],
        "title": "Country"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeadCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\+?[\\d\\s()-]{7,20}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "business": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Business"
          },
          "country": {
            "$ref": "#/components/schemas/Country"
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/Service"
            },
            "type": "array",
            "minItems": 1,
            "title": "Services"
          },
          "budget_range": {
            "$ref": "#/components/schemas/BudgetRange"
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Message"
          },
          "source_page": {
            "type": "string",
            "maxLength": 200,
            "pattern": "^/",
            "title": "Source Page"
          },
          "idempotency_key": {
            "type": "string",
            "format": "uuid",
            "title": "Idempotency Key"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website",
            "description": "Honeypot; must be empty or absent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "email",
          "country",
          "services",
          "budget_range",
          "message",
          "source_page",
          "idempotency_key"
        ],
        "title": "LeadCreate",
        "description": "Public inquiry payload. Validation here is the source of truth (research R8)."
      },
      "LeadCreated": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/LeadStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "post_process_token": {
            "type": "string",
            "title": "Post Process Token"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at",
          "post_process_token"
        ],
        "title": "LeadCreated"
      },
      "LeadEventOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "from_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LeadStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "to_status": {
            "$ref": "#/components/schemas/LeadStatus"
          },
          "actor_name": {
            "type": "string",
            "title": "Actor Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "from_status",
          "to_status",
          "actor_name",
          "created_at"
        ],
        "title": "LeadEventOut"
      },
      "LeadList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LeadOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "LeadList"
      },
      "LeadNoteOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "author_name": {
            "type": "string",
            "title": "Author Name"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "author_name",
          "body",
          "created_at"
        ],
        "title": "LeadNoteOut"
      },
      "LeadOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "business": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business"
          },
          "country": {
            "$ref": "#/components/schemas/Country"
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/Service"
            },
            "type": "array",
            "title": "Services"
          },
          "budget_range": {
            "$ref": "#/components/schemas/BudgetRange"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "$ref": "#/components/schemas/LeadStatus"
          },
          "source_page": {
            "type": "string",
            "title": "Source Page"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/LeadNoteOut"
            },
            "type": "array",
            "title": "Notes"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/LeadEventOut"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "business",
          "country",
          "services",
          "budget_range",
          "message",
          "status",
          "source_page",
          "created_at",
          "notes",
          "events"
        ],
        "title": "LeadOut"
      },
      "LeadStatus": {
        "type": "string",
        "enum": [
          "new",
          "contacted",
          "proposal_sent",
          "won",
          "lost"
        ],
        "title": "LeadStatus"
      },
...
      "NewNote": {
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "NewNote"
      },
...
      "Service": {
        "type": "string",
        "enum": [
          "digital_marketing",
          "creative_design",
          "web_software",
          "ui_ux_design",
          "video_content",
          "public_relations",
          "branding",
          "copywriting",
          "ai_automation",
          "iot_smart"
        ],
        "title": "Service",
        "description": "The agency's services, as listed in the brochure (ABOUT BUZZ CREW.pdf)."
      },
...
      "StatusChange": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/LeadStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusChange"
      },
      "SubscribeRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "title": "Email"
          },
          "source_page": {
            "type": "string",
            "maxLength": 200,
            "pattern": "^/",
            "title": "Source Page",
            "default": "/"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "SubscribeRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}